Refactored AuthProvider path
