+++ pages/dashboard.tsx
@@ -6,7 +6,7 @@ import nookies from "nookies";
 import { getAuth } from "firebase/auth";
 import { firebaseClient } from "../firebase/firebaseClient";
 import { firebaseAdmin } from "../firebase/firebaseAdmin";
-import { useAuth } from "../lib/auth/auth";
+import { useAuth } from "../lib/auth";
 
 const Dashboard: NextPageWithLayout<any> = () => {
   const user = useAuth()
